--- crates/dark-mode-daemon-macos/Cargo.toml
@@ -1,22 +1,0 @@
-[package]
-name = "dark-mode-daemon-macos"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-block2 = "0.5.1"
-dark-mode-daemon = { path = "../dark-mode-daemon" }
-objc2 = { version = "0.5.2" }
-objc2-app-kit = { version = "0.2.2", features = [
-  "NSApplication",
-  "NSResponder",
-] }
-objc2-foundation = { version = "0.2.2", features = [
-  "NSNotification",
-  "NSDistributedNotificationCenter",
-  "NSOperation",
-  "NSString",
-  "NSThread",
-  "NSUserDefaults",
-  "block2",
-] }
