+++ src/platform/mod.rs
@@ -1,0 +1,5 @@
+#[cfg(target_os = "macos")]
+pub mod macos;
+
+#[cfg(target_os = "linux")]
+pub mod linux;
